// app/models/user.ts
export class User {
    username: string;
    password: string;
    firstname: string;
    lastname: string;
    name: string;
    thumbnail: string;
    phone: string;
    street: string;
    city: string;
    state: string;
    postcode: string;
    gender: string;
    title: string;
     _id: string;
    socketid: string;
    selected:boolean;
    unreadMsgCount:number=0;
    lastMessage: string;
    msgDate: Date;
}